Assorted fixes and improvements for SGE
- Removed unused import
- Moved DRMAA extensions into refinement file: given that the DRMAA file is still almost pristine 3rd party code, except for a FFI vs Fiddle change
- Preferred using attr_readers to direct member calls
- Fixed error with @begin being used instead of @bin
- Cherry picked job id parser from job_array branch
